requirements.txt
Flask==1.1.1
gunicorn==20.0.4
pyTelegramBotAPI==3.6.7
requests==2.22.0
urllib3==1.25.7
beautifulsoup4==4.8.1
fake-useragent==0.1.11
python-dotenv
